+++ octis/models/ETM_model/data.py
@@ -1,24 +1,13 @@
-import os
-import random
-import pickle
 import numpy as np
 import torch
-import scipy.io
 
-def get_batch(tokens, counts, ind, vocab_size, emsize, device):
+def get_batch(tokens, counts, ind, vocab_size, device):
     """fetch input data by batch."""
     batch_size = len(ind)
     data_batch = np.zeros((batch_size, vocab_size))
     for i, doc_id in enumerate(ind):
         doc = tokens[doc_id]
         count = counts[doc_id]
-        #L = count.shape[1]
-        if len(doc) == 1:
-            doc = [doc.squeeze()]
-            count = [count.squeeze()]
-        else:
-            doc = doc#.squeeze()
-            count = count#.squeeze()
         if doc_id != -1:
             for j, word in enumerate(doc):
                 data_batch[i, word] = count[j]