+++ tests/test_octis.py
@@ -14,6 +14,7 @@ from octis.models.CTM import CTM
 from octis.models.NMF import NMF
 from octis.models.NMF_scikit import NMF_scikit
 from octis.models.ProdLDA import ProdLDA
+from octis.preprocessing.preprocessing import Preprocessing
 
 import numpy as np
 import os
@@ -27,6 +28,11 @@ def root_dir():
 @pytest.fixture
 def data_dir(root_dir):
     return root_dir + "/../preprocessed_datasets/"
+
+
+@pytest.fixture
+def embeddings_dir(root_dir):
+    return root_dir + "/../trained_embeddings/"
 
 
 def test_model_output_lda(data_dir):
@@ -82,6 +88,36 @@ def test_model_output_etm(data_dir):
     assert type(output['test-topic-document-matrix']) == np.ndarray
     assert output['test-topic-document-matrix'].shape == (num_topics, len(dataset.get_partitioned_corpus()[2]))
 
+def test_model_output_etm_with_corpus_containing_single_word_document(data_dir):
+    texts_path = data_dir+"/sample_texts/unprepr_docs.txt"
+    p = Preprocessing(vocabulary=None, max_features=None, remove_punctuation=True,
+                       lemmatize=False, stopword_list='english')
+    dataset = p.preprocess_dataset(
+        documents_path=texts_path,
+    )
+    num_topics = 3
+    model = ETM(num_topics=num_topics, num_epochs=5)
+    output = model.train_model(dataset)
+    assert 'topics' in output.keys()
+    assert 'topic-word-matrix' in output.keys()
+    assert 'test-topic-document-matrix' in output.keys()
+
+    # check topics format
+    assert type(output['topics']) == list
+    assert len(output['topics']) == num_topics
+
+    # check topic-word-matrix format
+    assert type(output['topic-word-matrix']) == np.ndarray
+    assert output['topic-word-matrix'].shape == (num_topics, len(dataset.get_vocabulary()))
+
+    # check topic-document-matrix format
+    assert type(output['topic-document-matrix']) == np.ndarray
+    assert output['topic-document-matrix'].shape == (num_topics, len(dataset.get_partitioned_corpus()[0]))
+
+    # check test-topic-document-matrix format
+    assert type(output['test-topic-document-matrix']) == np.ndarray
+    assert output['test-topic-document-matrix'].shape == (num_topics, len(dataset.get_partitioned_corpus()[2]))
+
 
 def test_model_output_etm_not_partitioned(data_dir):
     dataset = Dataset()
@@ -105,6 +141,143 @@ def test_model_output_etm_not_partitioned(data_dir):
     assert type(output['topic-document-matrix']) == np.ndarray
     assert output['topic-document-matrix'].shape == (num_topics, len(dataset.get_corpus()))
 
+
+def test_model_output_etm_with_pickle_embeddings_file(data_dir, embeddings_dir):
+    dataset = Dataset()
+    dataset.load_custom_dataset_from_folder(data_dir + '/M10')
+    num_topics = 3
+    model = ETM(num_topics=num_topics, num_epochs=5, train_embeddings=False, 
+        embeddings_path=embeddings_dir +'/test_example/example.pickle')
+    output = model.train_model(dataset)
+    assert 'topics' in output.keys()
+    assert 'topic-word-matrix' in output.keys()
+    assert 'test-topic-document-matrix' in output.keys()
+
+    # check topics format
+    assert type(output['topics']) == list
+    assert len(output['topics']) == num_topics
+
+    # check topic-word-matrix format
+    assert type(output['topic-word-matrix']) == np.ndarray
+    assert output['topic-word-matrix'].shape == (num_topics, len(dataset.get_vocabulary()))
+
+    # check topic-document-matrix format
+    assert type(output['topic-document-matrix']) == np.ndarray
+    assert output['topic-document-matrix'].shape == (num_topics, len(dataset.get_partitioned_corpus()[0]))
+
+    # check test-topic-document-matrix format
+    assert type(output['test-topic-document-matrix']) == np.ndarray
+    assert output['test-topic-document-matrix'].shape == (num_topics, len(dataset.get_partitioned_corpus()[2]))
+
+def test_model_output_etm_with_binary_word2vec_embeddings_file(data_dir, embeddings_dir):
+    dataset = Dataset()
+    dataset.load_custom_dataset_from_folder(data_dir + '/M10')
+    num_topics = 3
+    model = ETM(num_topics=num_topics, num_epochs=5, train_embeddings=False, 
+        embeddings_type='word2vec', embeddings_path=embeddings_dir +'/test_example/example.bin')
+    output = model.train_model(dataset)
+    assert 'topics' in output.keys()
+    assert 'topic-word-matrix' in output.keys()
+    assert 'test-topic-document-matrix' in output.keys()
+
+    # check topics format
+    assert type(output['topics']) == list
+    assert len(output['topics']) == num_topics
+
+    # check topic-word-matrix format
+    assert type(output['topic-word-matrix']) == np.ndarray
+    assert output['topic-word-matrix'].shape == (num_topics, len(dataset.get_vocabulary()))
+
+    # check topic-document-matrix format
+    assert type(output['topic-document-matrix']) == np.ndarray
+    assert output['topic-document-matrix'].shape == (num_topics, len(dataset.get_partitioned_corpus()[0]))
+
+    # check test-topic-document-matrix format
+    assert type(output['test-topic-document-matrix']) == np.ndarray
+    assert output['test-topic-document-matrix'].shape == (num_topics, len(dataset.get_partitioned_corpus()[2]))
+
+def test_model_output_etm_with_text_word2vec_embeddings_file(data_dir, embeddings_dir):
+    dataset = Dataset()
+    dataset.load_custom_dataset_from_folder(data_dir + '/M10')
+    num_topics = 3
+    model = ETM(num_topics=num_topics, num_epochs=5, train_embeddings=False, 
+        embeddings_type='word2vec', embeddings_path=embeddings_dir +'/test_example/example.txt', 
+        binary_embeddings=False)
+    output = model.train_model(dataset)
+    assert 'topics' in output.keys()
+    assert 'topic-word-matrix' in output.keys()
+    assert 'test-topic-document-matrix' in output.keys()
+
+    # check topics format
+    assert type(output['topics']) == list
+    assert len(output['topics']) == num_topics
+
+    # check topic-word-matrix format
+    assert type(output['topic-word-matrix']) == np.ndarray
+    assert output['topic-word-matrix'].shape == (num_topics, len(dataset.get_vocabulary()))
+
+    # check topic-document-matrix format
+    assert type(output['topic-document-matrix']) == np.ndarray
+    assert output['topic-document-matrix'].shape == (num_topics, len(dataset.get_partitioned_corpus()[0]))
+
+    # check test-topic-document-matrix format
+    assert type(output['test-topic-document-matrix']) == np.ndarray
+    assert output['test-topic-document-matrix'].shape == (num_topics, len(dataset.get_partitioned_corpus()[2]))
+
+def test_model_output_etm_with_headerless_text_word2vec_embeddings_file(data_dir, embeddings_dir):
+    dataset = Dataset()
+    dataset.load_custom_dataset_from_folder(data_dir + '/M10')
+    num_topics = 3
+    model = ETM(num_topics=num_topics, num_epochs=5, train_embeddings=False, 
+        embeddings_type='word2vec', embeddings_path=embeddings_dir +'/test_example/headerless_example.txt', 
+        binary_embeddings=False, headerless_embeddings=True)
+    output = model.train_model(dataset)
+    assert 'topics' in output.keys()
+    assert 'topic-word-matrix' in output.keys()
+    assert 'test-topic-document-matrix' in output.keys()
+
+    # check topics format
+    assert type(output['topics']) == list
+    assert len(output['topics']) == num_topics
+
+    # check topic-word-matrix format
+    assert type(output['topic-word-matrix']) == np.ndarray
+    assert output['topic-word-matrix'].shape == (num_topics, len(dataset.get_vocabulary()))
+
+    # check topic-document-matrix format
+    assert type(output['topic-document-matrix']) == np.ndarray
+    assert output['topic-document-matrix'].shape == (num_topics, len(dataset.get_partitioned_corpus()[0]))
+
+    # check test-topic-document-matrix format
+    assert type(output['test-topic-document-matrix']) == np.ndarray
+    assert output['test-topic-document-matrix'].shape == (num_topics, len(dataset.get_partitioned_corpus()[2]))
+
+def test_model_output_etm_with_keyedvectors_embeddings_file(data_dir, embeddings_dir):
+    dataset = Dataset()
+    dataset.load_custom_dataset_from_folder(data_dir + '/M10')
+    num_topics = 3
+    model = ETM(num_topics=num_topics, num_epochs=5, train_embeddings=False, 
+        embeddings_type='keyedvectors', embeddings_path=embeddings_dir +'/test_example/example.keyedvectors')
+    output = model.train_model(dataset)
+    assert 'topics' in output.keys()
+    assert 'topic-word-matrix' in output.keys()
+    assert 'test-topic-document-matrix' in output.keys()
+
+    # check topics format
+    assert type(output['topics']) == list
+    assert len(output['topics']) == num_topics
+
+    # check topic-word-matrix format
+    assert type(output['topic-word-matrix']) == np.ndarray
+    assert output['topic-word-matrix'].shape == (num_topics, len(dataset.get_vocabulary()))
+
+    # check topic-document-matrix format
+    assert type(output['topic-document-matrix']) == np.ndarray
+    assert output['topic-document-matrix'].shape == (num_topics, len(dataset.get_partitioned_corpus()[0]))
+
+    # check test-topic-document-matrix format
+    assert type(output['test-topic-document-matrix']) == np.ndarray
+    assert output['test-topic-document-matrix'].shape == (num_topics, len(dataset.get_partitioned_corpus()[2]))
 
 def test_model_output_lda_tomotopy(data_dir):
     dataset = Dataset()